requirements.txt
alembic==1.12.1
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
Flask==2.2.5
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.4
Mako==1.2.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyasn1==0.5.1
python-dotenv==0.21.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0